Move route configuration back to src/pages
