+++ src/pages/UploadDataPage.js
@@ -1,0 +1,10 @@
+import { UploadData } from 'src/workspace-data/upload-data/UploadData';
+
+export const navPaths = [
+  {
+    name: 'upload',
+    path: '/upload',
+    component: UploadData,
+    title: 'Upload',
+  },
+];
